requirements.txt
alembic==0.9.9
aniso8601==9.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==6.7
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography==36.0.1
cycler==0.11.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
fonttools==4.29.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.11.3
idna==2.6
importlib-metadata==4.11.2
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.3.2
lazy==1.3
Mako==1.0.7
Markdown==3.3.6
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
matplotlib==3.5.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
oauth2==1.9.0.post1
oauthlib==2.0.7
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pycparser==2.21
PyJWT==1.4.2
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-editor==1.0.3
pytz==2021.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.16.0
SQLAlchemy==1.3.0
SQLAlchemy-Utils==0.33.2
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
zipp==3.7.0
